+++ lib/include/bave/font/font.hpp
@@ -26,6 +26,11 @@ class Font {
 	/// \returns true on success.
 	auto load_from_bytes(std::vector<std::byte> file_bytes, float scale = scale_v) -> bool;
 
+	/// \brief Create font atlas for a specific text height.
+	/// \param height TextHeight to create atlas for.
+	/// \returns true if successful.
+	auto create_font_atlas(TextHeight height) -> bool;
+
 	[[nodiscard]] auto glyph_for(TextHeight height, Codepoint codepoint) -> Glyph;
 	[[nodiscard]] auto get_texture(TextHeight height) -> std::shared_ptr<Texture const>;